Simplify "add" command and support magnet links
If the argument is not a file in the local filesystem, the string
is passed to transmission. Errors are reported as transmission
reports them to us. This makes it possible to handle any weird
shit the Bittorrent community invents, like magnet links etc.

Also, the required version for Transmission::Client is bumped, as
the new upstream version correctly handles base64 encoding the
metainfo parameter. On the plus side, the requirement for
Regexp::Common is now dropped.