Mark files as wanted/not wanted in torrent view
Two new subcommands to the "torrent files" command are introduced:

 torrent <id> files on <pattern>
 torrent <id> files off <pattern>

The pattern is a regular expression; each file in the torrent
gets matched against the regexp. Files that match gets set to on
or off.

Actually, a third subcommand is introduced as well:

 torrent <id> files show

But just issuing "torrent <id> files" is a shortform of the
above, and will result in the same behavior as before.