Represent sizes as (at least) KiB
When a rate is around 1KiB/s, the rate can "jump" between 1KiB/s and 1000B/s
frequently. This can be confusing as you have to take care to notice the SI
prefix change. With this change all sizes above 0 are presented with an SI
prefix of at least kilo. 0 is still presented without any SI prefix: 0.00B.