Show how much you've downloaded for complete torrents
But only if you haven't downloaded as much as the seeded size. This
indicates that you have supplied data for the torrent by other means
than leeching, e.g. if you were the original uploader or just happened
to have the data already.

Some half assed refactoring is also included in this commit, for your
pleasure.