list: Greatly improve grep command
The list command has the -1 flag to get the torrents printed more
compactly, one line per torrent. There was code to support this
also for grep, but unfortunately it was severely broken.

  grep -1 foo

would result in a call to _list(..., "foo"), and _list would try to
use that as a numeric id. Now, not only does that work, you can also
limit the search to a range of torrents:

  grep foo 0-50

or combine them of course...